+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - minify_html plugin: Disable js and css minification by default
   unless `.css` and `.js` extensions are added.
 - Vento filters in components (ex: `{{ comp.button() |> toUpperCase }})
+- Slugifier: ignore errors for malformed URLs.
 - Updated dependencies: `sass`, `lightningcss`, `std`, `terser`,  `satori`, `decap-cms`, `ssx`, `tailwindcss`, `unocss` and some icons.
 
 ## [3.0.1] - 2025-05-10